+++ fastly/botmanagement/v1/doc.go
@@ -1,0 +1,3 @@
+// Package v1 contains functionality which configures ContentGuard,
+// controlling how Fastly responds to bot traffic.
+package v1