+++ fastly/botmanagement/v1/workspaces/doc.go
@@ -1,0 +1,2 @@
+// Package workspaces is the container for bot management settings and policies.
+package workspaces